project_teammates/requirements.txt
Django==2.2
psycopg2==2.8.2
djangorestframework==3.9.2
django-rest-auth==0.9.5
django-allauth==0.39.1
django-extensions==2.1.6
django-cors-headers==2.5.2